% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-S3-methods.R
\name{coef.dorem}
\alias{coef.dorem}
\title{S3 method for extracting model coefficients}
\usage{
\method{coef}{dorem}(object, ...)
}
\arguments{
\item{object}{Object of class \code{dorem}}

\item{...}{Extra arguments}
}
\description{
S3 method for extracting model coefficients
}
\examples{
data("bike_score")

banister_model <- dorem(
  Test_5min_Power ~ BikeScore,
  bike_score,
  method = "banister"
)
coef(banister_model)
}
