% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-package.R
\docType{package}
\name{dorem-package}
\alias{dorem-package}
\alias{_PACKAGE}
\title{dorem: Dose Response Modeling}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Understanding the effects of training dose to training response
    (i.e., performance) is one of the major goals of sports science. 
    The goal of this package is to provide easy-to-use dose-response models
    commonly applied in understanding aforementioned relationships.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dorem.net}
  \item Report bugs at \url{https://github.com/mladenjovanovic/dorem/issues}
}

}
\author{
\strong{Maintainer}: Mladen Jovanovic \email{coach.mladen.jovanovic@gmail.com}

Other contributors:
\itemize{
  \item Benedict Stephens Hemingway \email{b.stephens-hemingway@rgu.ac.uk} [contributor]
}

}
\keyword{internal}
