% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-fit.R
\name{dorem}
\alias{dorem}
\alias{dorem.default}
\alias{dorem.data.frame}
\alias{dorem.matrix}
\alias{dorem.formula}
\alias{dorem.recipe}
\title{Fit a \code{dorem}}
\usage{
dorem(x, ...)

\method{dorem}{default}(x, ...)

\method{dorem}{data.frame}(x, y, ...)

\method{dorem}{matrix}(x, y, ...)

\method{dorem}{formula}(formula, data, ...)

\method{dorem}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{dorem} object.
}
\description{
\code{dorem()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- dorem(predictors, outcome)

# Formula interface
mod2 <- dorem(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- dorem(rec, mtcars)

}
