% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-control.R
\name{dorem_loss_func}
\alias{dorem_loss_func}
\title{Default loss function}
\usage{
dorem_loss_func(obs, pred, weights, na.rm)
}
\arguments{
\item{obs}{Numeric vector. Observed values}

\item{pred}{Numeric vector. Predicted values}

\item{weights}{Numeric vector. Weights for weighting residuals}

\item{na.rm}{Logical}
}
\value{
Mean Squared Error (MSE)
}
\description{
This function is used inside the objective function to return model fit.
Default loss function is MSE.
}
\examples{
dorem_loss_func(seq(1:10), 5, rep(1, 10), na.rm = TRUE)
}
