% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-control.R
\name{dorem_no_link_func}
\alias{dorem_no_link_func}
\title{Default link function}
\usage{
dorem_no_link_func(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
By default there is no link function.
}
\examples{
dorem_no_link_func(1:10)
}
