% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-control.R
\name{dorem_perf_func}
\alias{dorem_perf_func}
\title{Default performance function}
\usage{
dorem_perf_func(obs, pred, na.rm)
}
\arguments{
\item{obs}{Numeric vector. Observed values}

\item{pred}{Numeric vector. Predicted values}

\item{na.rm}{Logical}
}
\value{
Named vector
}
\description{
List of estimators used to quantify model fit
}
\examples{
dorem_perf_func(rnorm(10), rnorm(10), na.rm = TRUE)
}
