% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injury-data.R
\docType{data}
\name{injury_data}
\alias{injury_data}
\title{Sample soft tissues injury data}
\format{Data frame with 7 variables and 658 observations:
\describe{
\item{Day}{Integer vector}
\item{Injury}{Integer. 0 indicated no-injury in the next 7 days or less, and
1 indicates injury in 7 days or less}
\item{Location}{Location of the injury}
\item{Type}{Type of injury}
\item{RPE.Load}{Numeric vector. RPE load}
\item{Total.Distance}{Numeric vector. Total running distance covered}
\item{HS.Distance}{Numeric vector. Hight Speed distance covered}
}}
\usage{
data(injury_data)
}
\description{
Sample soft tissues injury data
}
\keyword{datasets}
