% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-S3-methods.R
\name{plot.dorem}
\alias{plot.dorem}
\title{S3 method for plotting model}
\usage{
\method{plot}{dorem}(x, type = "pred", ...)
}
\arguments{
\item{x}{Object of class \code{dorem}}

\item{type}{Type of plot. Options are "pred", "coef", "perf", which "pred" being default}

\item{...}{Extra arguments}
}
\description{
S3 method for plotting model
}
\examples{
data("bike_score")

banister_model <- dorem(
  Test_5min_Power ~ BikeScore,
  bike_score,
  method = "banister",
  control = dorem_control(
        cv_folds = 3,
        cv_repeats = 5,
        shuffle = TRUE
  )
)
plot(banister_model)
}
