% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-predict.R
\name{predict.dorem}
\alias{predict.dorem}
\title{Predict from a \code{dorem}}
\usage{
\method{predict}{dorem}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{dorem} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{dorem}
}
\examples{
train <- mtcars[1:20,]
test <- mtcars[21:32, -1]

# Fit
mod <- dorem(mpg ~ cyl + log(drat), train)

# Predict, with preprocessing
predict(mod, test)

}
