% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-S3-methods.R
\name{print.dorem}
\alias{print.dorem}
\title{S3 method for printing model results}
\usage{
\method{print}{dorem}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dorem}}

\item{...}{Extra arguments}
}
\description{
S3 method for printing model results
}
\examples{
data("bike_score")

banister_model <- dorem(
  Test_5min_Power ~ BikeScore,
  bike_score,
  method = "banister",
  control = dorem_control(
        cv_folds = 3,
        cv_repeats = 5,
        shuffle = TRUE
  )
)
print(banister_model)
}
