% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-package.R
\docType{package}
\name{dorem-package}
\alias{dorem}
\alias{dorem-package}
\title{dorem: Dose Response Modeling}
\description{
This package models training dose to training response.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mladenjovanovic/dorem}
  \item Report bugs at \url{https://github.com/mladenjovanovic/dorem/issues}
}

}
\author{
\strong{Maintainer}: Mladen Jovanovic \email{coach.mladen.jovanovic@gmail.com}

}
\keyword{internal}
