% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-fit.R
\name{dorem}
\alias{dorem}
\alias{dorem.default}
\alias{dorem.data.frame}
\alias{dorem.matrix}
\alias{dorem.formula}
\alias{dorem.recipe}
\title{Fit a \code{dorem}}
\usage{
dorem(x, ...)

\method{dorem}{default}(x, ...)

\method{dorem}{data.frame}(x, y, ...)

\method{dorem}{matrix}(x, y, ...)

\method{dorem}{formula}(formula, data, ...)

\method{dorem}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{dorem} object.
}
\description{
\code{dorem()} fits a model.
}
\examples{
require(tidyverse)

data("bike_score")

banister_model <- dorem(
  Test_5min_Power ~ BikeScore,
  bike_score,
  method = "banister"
)

bike_score$pred <- predict(banister_model, bike_score)$.pred

ggplot(bike_score, aes(x = Day, y = pred)) +
  theme_bw() +
  geom_line() +
  geom_point(aes(y = Test_5min_Power), color = "red") +
  ylab("Test 5min Power")

}
