% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-control.R
\name{dorem_control}
\alias{dorem_control}
\title{Constructor for the \code{dorem} control object}
\usage{
dorem_control(
  weights = NULL,
  na.rm = TRUE,
  loss_func = dorem_loss_func,
  link_func = dorem_no_link_func,
  perf_func = dorem_perf_func,
  optim_method = valid_optimization_methods(),
  optim_maxit = 1000,
  optim_VTR = -Inf,
  optim_grid_n = 5,
  optim_trace = FALSE,
  coefs_start = NULL,
  coefs_lower = NULL,
  coefs_upper = NULL,
  cv_repeats = NULL,
  cv_folds = NULL,
  dose_before_response = TRUE,
  shuffle = FALSE,
  iter = TRUE,
  seed = NULL
)
}
\arguments{
\item{weights}{Default is NULL}

\item{na.rm}{Default is TRUE}

\item{loss_func}{Function used in optimization objective function. Default is \code{\link{dorem_loss_func}}
which returns MSE}

\item{link_func}{Function used to convert predictions. Default is \code{\link{dorem_no_link_func}}
which returns the original value}

\item{perf_func}{Function used to quantify model fit. Default is \code{\link{dorem_perf_func}}}

\item{optim_method}{Optimization method. The list of supported methods is \code{c("L-BFGS-B", "DE", "CMA-ES", "gridSearch")}}

\item{optim_maxit}{Maximal number of iterations for the optimization method. Default is 1000}

\item{optim_VTR}{Value To Reach in the optimization method. Default is -Inf}

\item{optim_grid_n}{Number of search values for each predictor used in gridSearch method. Default is 5}

\item{optim_trace}{Should optimization trace be shown? Default is FALSE}

\item{coefs_start}{Starting values for coefficients}

\item{coefs_lower}{Lower bound for coefficients}

\item{coefs_upper}{Upper bound for coefficients}

\item{cv_repeats}{Number of CV repeats}

\item{cv_folds}{Number of CV folds}

\item{dose_before_response}{If dose variable and performance variable happen on the same row (i.e. same day),
does dose precedes response? This assumes that the dose is done before response. Default is \code{TRUE}}

\item{shuffle}{Should shuffle be performed? Default is FALSE}

\item{iter}{Should iter be shown? Default is TRUE}

\item{seed}{Random number seed}
}
\description{
Constructor for the \code{dorem} control object
}
\examples{
data("bike_score")

banister_model <- dorem(
  Test_5min_Power ~ BikeScore,
  bike_score,
  method = "banister",
  control = dorem_control(
        cv_folds = 3,
        cv_repeats = 5,
        shuffle = TRUE
  )
)
}
