% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-control.R
\name{dorem_control}
\alias{dorem_control}
\title{Constructor for dorem control object}
\usage{
dorem_control(
  loss_func = function(obs, pred, weights, na.rm = TRUE) {     mean(weights * ((pred -
    obs))^2, na.rm = na.rm) },
  link_func = function(x) {     x },
  perf_func = function(obs, pred, na.rm = TRUE) {     list(RMSE = sqrt(mean((pred -
    obs)^2, na.rm = na.rm)), MAE = mean(abs(pred - obs), na.rm = na.rm), maxAbsErr =
    max(abs(pred - obs), na.rm = na.rm), MAPE = 100 * mean(abs((pred - obs)/obs), na.rm =
    na.rm)) },
  optim_method = ifelse(any(!is.null(coefs_lower), !is.null(coefs_upper)), "L-BFGS-B",
    "BFGS"),
  optim_maxit = 1000,
  optim_trace = FALSE,
  coefs_start = NULL,
  coefs_lower = NULL,
  coefs_upper = NULL,
  cv_repeats = NULL,
  cv_folds = NULL,
  iter = FALSE,
  seed = round(stats::runif(1, 1, 10000))
)
}
\arguments{
\item{loss_func}{Will be explained}

\item{link_func}{Will be explained}

\item{perf_func}{Will be explained}

\item{optim_method}{Will be explained}

\item{optim_maxit}{Will be explained}

\item{optim_trace}{Will be explained}

\item{coefs_start}{Will be explained}

\item{coefs_lower}{Will be explained}

\item{coefs_upper}{Will be explained}

\item{cv_repeats}{Will be explained}

\item{cv_folds}{Will be explained}

\item{iter}{Will be explained}

\item{seed}{Will be explained}
}
\description{
Constructor for dorem control object
}
