% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorem-predict.R
\name{predict.dorem}
\alias{predict.dorem}
\title{Predict from a \code{dorem}}
\usage{
\method{predict}{dorem}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{dorem} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{dorem}
}
\examples{
require(tidyverse)

data("bike_score")

banister_model <- dorem(
  Test_5min_Power ~ BikeScore,
  bike_score,
  method = "banister"
)

bike_score$pred <- predict(banister_model, bike_score)$.pred

ggplot(bike_score, aes(x = Day, y = pred)) +
  theme_bw() +
  geom_line() +
  geom_point(aes(y = Test_5min_Power), color = "red") +
  ylab("Test 5min Power")
}
